refactor: improve file organization in orders service
