+++ orders-service/src/domain/order.ts
@@ -1,25 +1,10 @@
+import { InsertOrder } from "../repository/insert-order"
+
 /**
  * this interface represents an Order
  */
 export interface Order extends InsertOrder {
 	_id: string
-}
-
-/**
- * this interface represents the information needed to insert an order into the repository
- */
-export interface InsertOrder extends NewOrder {
-	state: OrderState,
-}
-
-/**
- * this interface represents the information needed to create a new pending order
- */
-export interface NewOrder {
-	customerEmail: string,
-	price: number,
-	type: OrderType,
-	items: OrderItem[]
 }
 
 /**
