+++ orders-service/src/repository/insert-order.ts
@@ -1,0 +1,11 @@
+import { OrderState } from "../domain/order";
+import { NewOrder } from "../routes/new-order";
+
+/**
+ * this interface represents the information needed to insert an order into the repository
+ */
+
+export interface InsertOrder extends NewOrder {
+	state: OrderState;
+}
+
