+++ orders-service/src/routes/new-order.ts
@@ -1,0 +1,11 @@
+import { OrderType, OrderItem } from "../domain/order";
+
+/**
+ * this interface represents the information needed to create a new pending order
+ */
+export interface NewOrder {
+	customerEmail: string;
+	price: number;
+	type: OrderType;
+	items: OrderItem[];
+}
